Se implementa la lógica de actualización de eventos de conejos de la Granja a través del botón de Guardado

+++ CuniculappAndroid/app/src/main/java/upv/cuniculappteam/cuniculapp/model/labors/Sale.java
@@ -55,7 +55,7 @@ public class Sale extends Traceable
         return averageWeight;
     }
 
-    public void setAverageWeight(Integer averageWeight) {
+    public void setAverageWeight(float averageWeight) {
         this.averageWeight = averageWeight;
     }
